+++ dwtools/HMM.h
@@ -43,11 +43,11 @@ Collection_define (HMMObservationList, OrderedOf, HMMObservation) {
 /********** class HMMBaumWelch **********/
 
 Thing_define (HMMBaumWelch, Daata) {
-	long capacity;
-	long numberOfTimes;
-	long totalNumberOfSequences;
-	long numberOfStates;
-	long numberOfSymbols;
+	integer capacity;
+	integer numberOfTimes;
+	integer totalNumberOfSequences;
+	integer numberOfStates;
+	integer numberOfSymbols;
 	double lnProb;
 	double minProb;
 	double **alpha;
@@ -72,9 +72,9 @@ Thing_define (HMMObservationSequence, Table) {
 Collection_define (HMMObservationSequenceBag, CollectionOf, HMMObservationSequence) {
 };
 
-autoHMMObservationSequence HMMObservationSequence_create (long numberOfItems, long dataLength);
+autoHMMObservationSequence HMMObservationSequence_create (integer numberOfItems, integer dataLength);
 
-void HMMObservationSequence_removeObservation (HMMObservationSequence me, long index);
+void HMMObservationSequence_removeObservation (HMMObservationSequence me, integer index);
 
 autoStrings HMMObservationSequence_to_Strings (HMMObservationSequence me);
 
@@ -88,32 +88,32 @@ autoTableOfReal HMMObservationSequence_to_TableOfReal_transitions (HMMObservatio
 
 autoTableOfReal HMM_and_HMMObservationSequence_to_TableOfReal_transitions (HMM me, HMMObservationSequence thee, int probabilities);
 
-long HMMObservationSequence_getNumberOfObservations (HMMObservationSequence me);
+integer HMMObservationSequence_getNumberOfObservations (HMMObservationSequence me);
 
-long HMMObservationSequenceBag_getLongestSequence (HMMObservationSequenceBag me);
+integer HMMObservationSequenceBag_getLongestSequence (HMMObservationSequenceBag me);
 
 autoTableOfReal HMMStateSequence_to_TableOfReal_transitions (HMMStateSequence me);
 
 autoTableOfReal HMM_and_HMMStateSequence_to_TableOfReal_transitions (HMM me, HMMStateSequence thee, int probabilities);
 
-autoHMMStateSequence HMMStateSequence_create (long numberOfItems);
+autoHMMStateSequence HMMStateSequence_create (integer numberOfItems);
 
 autoStrings HMMStateSequence_to_Strings (HMMStateSequence me);
 
 void HMMState_setLabel (HMMState me, char32 *label);
 
 
-autoHMM HMM_create (int leftToRight, long numberOfStates, long numberOfObservationSymbols);
+autoHMM HMM_create (int leftToRight, integer numberOfStates, integer numberOfObservationSymbols);
 
 autoHMM HMM_createSimple (int leftToRight, const char32 *states_string, const char32 *symbols_string);
 
-autoHMM HMM_createContinuousModel (int leftToRight, long numberOfStates, long numberOfObservationSymbols,
-	long numberOfMixtureComponentsPerSymbol, long componentDimension, long componentStorage);
+autoHMM HMM_createContinuousModel (int leftToRight, integer numberOfStates, integer numberOfObservationSymbols,
+	integer numberOfMixtureComponentsPerSymbol, integer componentDimension, integer componentStorage);
 
-autoHMM HMM_createFullContinuousModel (int leftToRight, long numberOfStates, long numberOfObservationSymbols,
-	long numberOfFeatureStreams, long *dimensionOfStream, long *numberOfGaussiansforStream);
+autoHMM HMM_createFullContinuousModel (int leftToRight, integer numberOfStates, integer numberOfObservationSymbols,
+	integer numberOfFeatureStreams, integer *dimensionOfStream, integer *numberOfGaussiansforStream);
 
-autoHMM HMM_createFromHMMObservationSequence (HMMObservationSequence me, long numberOfStates, int leftToRight);
+autoHMM HMM_createFromHMMObservationSequence (HMMObservationSequence me, integer numberOfStates, int leftToRight);
 
 void HMM_draw (HMM me, Graphics g, int garnish);
 
@@ -147,22 +147,22 @@ void HMM_unExpandPCA (HMM me);
 /*
 	Set the probabilities. A probability zero value indicates that this p cannot be changed during training/learning.
 */
-void HMM_setTransitionProbabilities (HMM me, long state_number, char32 *state_probs);
+void HMM_setTransitionProbabilities (HMM me, integer state_number, char32 *state_probs);
 
-void HMM_setEmissionProbabilities (HMM me, long state_number, char32 *emission_probs);
+void HMM_setEmissionProbabilities (HMM me, integer state_number, char32 *emission_probs);
 
 void HMM_setStartProbabilities (HMM me, char32 *probs);
 
 
-double HMM_getProbabilityAtTimeBeingInState (HMM me, long itime, long istate);
+double HMM_getProbabilityAtTimeBeingInState (HMM me, integer itime, integer istate);
 
-double HMM_getProbabilityAtTimeBeingInStateEmittingSymbol (HMM me, long itime, long istate, long isymbol);
+double HMM_getProbabilityAtTimeBeingInStateEmittingSymbol (HMM me, integer itime, integer istate, integer isymbol);
 
-double HMM_getExpectedValueOfDurationInState (HMM me, long istate);
+double HMM_getExpectedValueOfDurationInState (HMM me, integer istate);
 
-double HMM_getProbabilityOfStayingInState (HMM me, long istate, long numberOfTimeUnits);
+double HMM_getProbabilityOfStayingInState (HMM me, integer istate, integer numberOfTimeUnits);
 
-double HMM_and_HMM_getCrossEntropy (HMM me, HMM thee, long observationLength, int symmetric);
+double HMM_and_HMM_getCrossEntropy (HMM me, HMM thee, integer observationLength, int symmetric);
 
 double HMM_and_HMM_and_HMMObservationSequence_getCrossEntropy (HMM me, HMM thee, HMMObservationSequence him);
 
@@ -172,7 +172,7 @@ autoTableOfReal HMM_extractEmissionProbabilities (HMM me);
 
 /* HMM & HMMObservationSequence ****/
 
-autoHMMObservationSequence HMM_to_HMMObservationSequence (HMM me, long initialState, long numberOfItems);
+autoHMMObservationSequence HMM_to_HMMObservationSequence (HMM me, integer initialState, integer numberOfItems);
 
 autoHMMStateSequence HMM_and_HMMObservationSequence_to_HMMStateSequence (HMM me, HMMObservationSequence thee);
